+++ ch6/exercises/test-VectorStack.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+g++ test-VectorStack.cpp -o test-VectorStack
+time ./test-VectorStack
+rm ./test-VectorStack